image_crawler/requirements.txt
Pillow
numpy
pymongo
